in debian/control: 	update descriptions 	remove paella-dbserver package (move the stanza to another file 		to be used later) update interactive.py to better match current code in databases/quickstart-vbox: 	updated default-environment 	added all_in_one filesystem 	added default_charset variable 	added some scripts for base trait (most don't do anything, I'm 		just using them to test things) 	added more stuff to base trait 	update debconf template in base trait in kde.trait.viewbrowser: 	get confirmation before deleting script in db.trait.main: 	make suite arg mandatory instead of optional keyword (most 		of the code should be calling it this way by now) in base.template: 	only print statements if DEBUG is set in TemplatedEnvironment in installer.base: 	add path to "string type" commands for runlog 	fix run_process_script to pass a list to runlog in installer.chroot: 	now mounting and umounting /dev/pts 	added some debug logging (will probably remove soon) 	use list instead of string for extract tarball command 	now we revert to debootstrap when a base tarball isn't located 	create new /etc/resolv.conf in ready_base_for_install in installer.trait: 	added new preseed process to process debconf templates 	raise an error if partial debs are detected in /var/cache/apt/archives/partial 	move install_debconf_template closer to bottom of class (we'll be 		removing this soon) 	setting debconf in the preseed process can cause an error if the 		question isn't yet in the database (but not having the 		preseed keeps some packages, like locales, from being 		properly configured.  We now check to see if there was an 		error with the preseed process, and set the selections again 		during the templates process, and also make sure that the 		setting of selections during the templates process is 		successful 	make a list for chroot script command 	override run_process_script to handle chroot script in installer.machine: 	forgot to add umount_target_devpts to DEFAULT_PROCESSES 	get logdirs from default-environment in installer.machinehelper: 	remove environment variables that setup_disk_fai set in installer.toolkit: 	major overhaul here 	still unfinished, but looking a lot better in installer.util.disk: 	setup_disk_fai now returns command instead of running it in setup.py: 	fix get_version to work right

+++ src/paella/kde/trait/viewbrowser.py
@@ -198,9 +198,12 @@ class TraitView(ViewBrowser):
             self.resetView()
         elif context == 'template':
             debug(context, ident)
-            template = self._convert_template_id(ident)
-            self.doc.trait.delete_template(template)
-            self.resetView()
+            ans = KMessageBox.questionYesNo(self,
+                                            "really delete this template?")
+            if ans == KMessageBox.Yes:
+                template = self._convert_template_id(ident)
+                self.doc.trait.delete_template(template)
+                self.resetView()
         else:
             raise RuntimeError, '%s context not implemented' % context
